Feature/contact lookup (#23)
# Description

Adds contact lookup and typeahead

This is in a place where someone can definitely start looking at it. I could use some help implementing the unfinished tasks, I'm running out of free time to work on it

# Notes
Currently this is implemented using the `AutoComplete` mui component. I've overridden the styles that display the search results in a dropdown as the user types. I implemented it this way first for testing purposes, but it might actually be useful in the future to use it because then we can overlay the search on any page, instead of having to navigate to a search results page. I wanted to leave it this way for now and have a discussion on that. Otherwise I'll just change it to a basic text input component

# Feature
Typeahead flow #6 

# Todo
- [x] Pull typeahead data and filter it to match user input
- [x] Display matched contact UIDs to user
- [x] Navigate to URL on click
- [x] Fix user icon bc it looks weird for some reason
- [x] Stick search component in app bar
- [x] Navigate to search results page when user enters search text
- [x] Clear button
- [x] Char count display
- [x] Validate search string user enters and display error text when it's formatted incorrectly
- [x] Remove contacts fetch and associated redux stuff
- [x] Display most recent user
    - this one is interesting bc even though we already have the info of the current user in state, we don't have their UID

~Allow user to create a new contact when they search for one that doesn't exist~ moved to #21 
~Format UID as user enters it (nice to have)~ moved to #26
~Add loading spinner (add to debounce issue)~ moved to #26
~Debounce on input (required, OR move to separate Issue)~ moved to #26