Adds a check box that disables the slider
There is now a value in this.state that decides whether or not the
checkbox is disabled by default. It works, but I think I'm breaking
React convention.

I'm getting this error:
proxyConsole.js:56 Warning: setState(...): Can only update a mounted or
mounting component. This usually means you called setState() on an
unmounted component.
This is a no-op. Please check the code for the TransitionGroup component.