Feature/intake (#36)
Updates since last request: state is updated when form elements are altered; many UI changes; added 'Return Visit' functionality; switched to date picker; added elements in reference to Excel doc; began work on clear form

Please refer to issues #31, #32, #33, #34, #35 for next steps

---

* getting contact info page firebase flow to start working, need to put the call in the right place and get some eyes on it

* WIP--show events and submit intake form

* intake page

* added a bulk of functionality to the intake form

* all onchange complete so far

* addressed most of evans comments, updated styling, and adjusted groupings of fields

* some general cleanup, fixing some routing in app.js, removing a number of contact related files taht are not being used and actually exist in ContactInfo/*

* alex review

* revert

* breaking out all of the forms into separate components. simplifying the
main form component to do nothing other than manage state, and pass
update/create functions to the sub-form components. some of the styles
have suffered in the process of refactoring it this way, but the data
flow is super clean now and the form is super modular and rather
extensible. ultimately I'd like to start pulling the form details from
the backend as opposed to storing all of the options/choices on the
client side. form data posts now and all of the excel items have been
added from the sheet

* forgot to make toggles and selects fully controlled and use value from props. setup event submission to only submit data for visible forms

* adds geolocation to form data collection if supported

* option picker looked really awful at different widths, finally got around to fixing it

* forgot to make slider fully controlled as well

* event strucutre

* changing name of periodic form, removing yarn lock

* conditionally display in care for hiv/hepc if the contact is positive

* PR Comments